resource: support for synchronously releasable resources.
This commit adds the possibility to configure a resource to be
synchronously releasable. For such resource the client must set
the release callback, which will be called when the resource
ownership needs to be transfered to another client and before
the current owner actually loses access to the resource.

In the release callback the current owner should stop using the
resource and release all actual physical/OS resources allocated
for it. Once this is done and the release callback returns, the
resource ownership transfer workflow continues as ordinary: the
resource owner is changed and the former and new owners receive
ownership change notifications.

Signed-off-by: Krisztian Litkey <kli@iki.fi>